% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candy.R
\docType{data}
\name{candy}
\alias{candy}
\title{Candy data}
\format{
Two matrices of dimension 18 x 6.
}
\usage{
data(candy)
}
\description{
Measurements from sensory analysis (professional tasting)
on a number of candy products obtained by sensory labs. The two labs
and the associated data sets are parts of a larger study described in Tomic et al. (2010),
}
\references{
Tomic, O., Luciano, G., Nilsen, A., Hyldig, G., Lorensen, K., Næs, T. (2010). 
Analysing sensory panel performance in a proficiency test using the PanelCheck software. 
European Food Research and Technology. 230. 3, 497-511
}
