% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixCorrelation.R
\docType{package}
\name{MatrixCorrelation}
\alias{MatrixCorrelation}
\alias{MatrixCorrelation-package}
\alias{MatrixCorrelation-package}
\title{Similiarity of Matrices Coefficients}
\description{
Computation and visualization of matrix correlation coefficients.
  The main method is the Similarity of Matrices Index, while various related
  measures like r1, r2, r3, r4, Yanai's GCD, RV, RV2 and adjusted RV are included
  for comparison.
}
\seealso{
\code{\link{SMI}}, \code{\link{plot.SMI}} (print.SMI/summary.SMI), \code{\link{RV}} (RV2/RVadj), \code{\link{r1}} (r2/r3/r4/GCD), \code{\link{allCorrelations}} (matrix correlation comparison).
}
\author{
\strong{Maintainer}: Kristian Hovde Liland \email{kristian.liland@nmbu.no}

Other contributors:
\itemize{
  \item Tormod Næs [contributor]
  \item Ulf Geir Indahl [contributor]
}

}
