% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMI.R
\name{SMI}
\alias{MatrixCorrelation}
\alias{SMI}
\title{Similarity of Matrices Index (SMI)}
\usage{
SMI(X1, X2, ncomp1 = Rank(X1) - 1, ncomp2 = Rank(X2) - 1,
  projection = "Orthogonal", Scores1 = NULL, Scores2 = NULL)
}
\arguments{
\item{X1}{first \code{matrix} to be compared (\code{data.frames} are also accepted).}

\item{X2}{second \code{matrix} to be compared (\code{data.frames} are also accepted).}

\item{ncomp1}{maximum number of subspace components from the first \code{matrix}.}

\item{ncomp2}{maximum number of subspace components from the second \code{matrix}.}

\item{projection}{type of projection to apply, defaults to "Orthogonal", alternatively "Procrustes".}

\item{Scores1}{user supplied score-\code{matrix} to replace singular value decomposition of first \code{matrix}.}

\item{Scores2}{user supplied score-\code{matrix} to replace singular value decomposition of second \code{matrix}.}
}
\value{
A matrix containing all combinations of components. Its class is "SMI" associated
with print, plot, summary methods.
}
\description{
Matrix similarity estimated through projections of stable subspaces.
}
\details{
For each combination of components SMI is computed based on the selected
 orhtogonal subspace and projection method. If the projection is orthogonal, complemental
 SMI values are computed for use in plotting and testing.
}
\examples{
X1  <- matrix(rnorm(100*300),100,300)
usv <- svd(X1)
X2  <- usv$u[,-3] \%*\% diag(usv$d[-3]) \%*\% t(usv$v[,-3])

(smi <- SMI(X1,X2,5,5))
plot(smi)

}
\author{
Kristian Hovde Liland
}
\references{
Similarity of Matrices Index - Ulf G. Indahl, Tormod Næs, Kristian Hovde Liland
}
\seealso{
\code{\link{plot.SMI}} (print.SMI/summary.SMI), \code{\link{RV}} (RV2/RVadj), \code{\link{r1}} (r2/r3/r4/GCD), \code{\link{allCorrelations}} (matrix correlation comparison).
}

